Rename all instances of cog to cogvlm
