+++ examples/cogvlm/cogvlm_client.py
