+++ inference/core/entities/responses/cogvlm.py
