+++ inference/models/cogvlm/__init__.py
@@ -1,1 +1,1 @@
-from inference.models.cogvlm.cog import CogVLM
+from inference.models.cogvlm.cogvlm import CogVLM
